requirements.txt
openai-whisper
gtts
pyaudio
sounddevice
numpy
scipy
google-generativeai
python-dotenv
playsound==1.2.2
colorama
